add swipe left, delete hit from list, save hit id, filter deletes items

+++ app/src/main/java/cl/orlandoormazabal/reigndemoapp/domain/RepoImp.kt
@@ -2,6 +2,7 @@ package cl.orlandoormazabal.reigndemoapp.domain
 
 import cl.orlandoormazabal.reigndemoapp.base.connection.NetworkState
 import cl.orlandoormazabal.reigndemoapp.data.model.Hit
+import cl.orlandoormazabal.reigndemoapp.data.model.HitIdEntity
 import cl.orlandoormazabal.reigndemoapp.extensions.toHitEntity
 
 class RepoImp(
@@ -13,15 +14,17 @@ class RepoImp(
     ): Repo {
 
     override suspend fun getHits(): List<Hit> {
-        return if(networkState.isInternetAvailable().isInternetAvailable) {
+         if (networkState.isInternetAvailable().isInternetAvailable) {
             hitDao.deletePreviousHits()
             val newHitList = remoteDataSource.getRemoteHits()
             newHitList.map { it.toHitEntity() }.forEach { entity ->
                 hitDao.insertHit(entity)
             }
-            newHitList
-        } else {
-            localDataSource.getLocalHits(hitDao)
         }
+        return localDataSource.getLocalHits(hitDao)
+    }
+
+    override suspend fun insertDeleteHitId(id: String) {
+        hitIdDao.insert(HitIdEntity(0, id))
     }
 }
